// types/database.d.ts
export type position = "bench" | "gk" | "def" | "mid" | "att";
export type forecast = "a" | "u" | "m";
// These are the types for the database
export interface users {
  id: number;
  username: string;
  password: string;
  throttle: number;
  active: boolean;
  inActiveDays: number;
  google: string;
  github: string;
  admin: boolean;
  favoriteLeague: number | null;
  theme: string | null;
  locale: string | null;
}
export interface players {
  uid: string;
  name: string;
  nameAscii: string;
  club: string;
  pictureID: number;
  value: number;
  sale_price: number;
  position: position;
  forecast: forecast;
  total_points: number;
  average_points: number;
  last_match: number;
  locked: boolean;
  exists: boolean;
  league: string;
}
// ...
export interface historicalPlayers {
  time: number;
  uid: string;
  name: string;
  nameAscii: string;
  club: string;
  pictureID: number;
  value: number;
  sale_price: number;
  position: position;
  forecast: forecast;
  total_points: number;
  average_points: number;
  last_match: number;
  exists: boolean;
  league: string;
}